+++ library/core/api/src/main/kotlin/de/lise/fluxflow/api/workflow/WorkflowDefinition.kt
@@ -9,4 +9,9 @@ interface WorkflowDefinition<TModel> {
      * A list of listeners to be informed, whenever there are changes to the step model.
      */
     val updateListeners: List<ModelListenerDefinition<TModel>>
+    
+    /**
+     * A simple key-value map providing additional metadata about this workflow definition.
+     */
+    val metadata: Map<String, Any>
 }