+++ library/core/engine/src/main/kotlin/de/lise/fluxflow/engine/workflow/WorkflowDefinitionImpl.kt
@@ -4,5 +4,6 @@ import de.lise.fluxflow.api.workflow.ModelListenerDefinition
 import de.lise.fluxflow.api.workflow.WorkflowDefinition
 
 class WorkflowDefinitionImpl<TModel>(
-    override val updateListeners: List<ModelListenerDefinition<TModel>>
+    override val updateListeners: List<ModelListenerDefinition<TModel>>,
+    override val metadata: Map<String, Any> = emptyMap()
 ) : WorkflowDefinition<TModel>